requirements.txt
-e ../acosmibot-core
python-magic
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiomysql==0.3.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.8.0
greenlet==3.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
mysql-connector-python==9.4.0
numpy==2.3.5
propcache==0.4.1
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
redis==7.1.0
requests==2.32.5
spglib==2.6.0
SQLAlchemy==2.0.45
stripe==14.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
utils==1.0.2
Werkzeug==3.1.3
xmltodict==1.0.2
yarl==1.22.0
